+++ src/pages/User/hooks/index.ts
@@ -1,4 +1,2 @@
 export * from './useInput';
-export * from './useLogin';
-export * from './useRegister';
 export * from './useAuthForm';
